+++ discord-app/src/discord_bot/message_create/MessageCreate.ts
@@ -1,67 +1,40 @@
+import { Message } from 'discord.js';
+import birthdayEventCommand from './birthdayEventCommand';
+import disconnectVoiceChannelCommand from './disconnectVoiceChannelCommand';
 import deleteReplyMessage from './deleteReplyMessage';
 import generateResponseForGemini from './generateResponseForGemini';
+import Gemini from '../../gemini/Gemini';
 import greetToNew235Member from './greetToNew235Member';
 import helpCommand from './helpCommand';
+import joinVoiceChannelCommand from './joinVoiceChannelCommand';
+import menEventCommand from './menEventCommand';
+import roomDivisionCommand from './roomDivisionCommand';
 import storeMessage from './storeMessage';
 import storeToNew235MemberBirthday from './storeToNew235MemberBirthday';
 import readTextForVoiceVox from './readTextForVoiceVox';
 import reactToBirthday235MemberMessage from './reactToBirthday235MemberMessage';
 import reactToBirthdayMillionMemberMessage from './reactToBirthdayMillionMemberMessage';
 import reactToUsedMaleEventCommandMessage from './reactToUsedMaleEventCommandMessage';
-
-const {
-  Client,
-  EmbedBuilder,
-  GuildMember,
-  GuildScheduledEventEntityType,
-  GuildScheduledEventManager,
-  GuildScheduledEventPrivacyLevel,
-  Message,
-} = require('discord.js');
-
-const { joinVoiceChannel } = require('@discordjs/voice');
-const fs = require('fs');
-const DiscordBot = require('../DiscordBot').default;
-const Gemini = require('../../gemini/Gemini').default;
-const VoiceVox = require('../../voice_vox/VoiceVox').default;
-const BirthdayFor235MemberRepository =
-  require('../../../repositories/BirthdayFor235MemberRepository').default;
-
-const { BirthdayFor235Member, BirthdayForMillionMember, DeleteMessage, DictWord } =
-  require('../../../models/index').default;
+import VoiceVox from '../../voice_vox/VoiceVox';
+import { DiscordBotType } from '../DiscordBotType';
 
 /**
  * メッセージが送信された時に行う処理クラス
  */
 export default class MessageCreate {
-  private discordBot: typeof DiscordBot;
+  private discordBot: DiscordBotType;
 
-  private gemini: typeof Gemini;
+  private gemini: Gemini;
 
-  private voiceVox: typeof VoiceVox;
+  private voiceVox: VoiceVox;
 
   private readonly prefix = '235';
 
-  private readonly setTimeoutSec = 15_000;
-
-  private readonly maleEventEmojiList = [
-    '1️⃣',
-    '2️⃣',
-    '3️⃣',
-    '4️⃣',
-    '5️⃣',
-    '6️⃣',
-    '7️⃣',
-    '8️⃣',
-    '9️⃣',
-    '🔟',
-  ];
-
   /**
-   * @param {DiscordBot} discordBot DiscordBotクラス
+   * @param {DiscordBotType} discordBot DiscordBotクラス
    * @param {VoiceVox} voiceVox VoiceVoxクラス
    */
-  constructor(discordBot: typeof DiscordBot, voiceVox: typeof VoiceVox) {
+  constructor(discordBot: DiscordBotType, voiceVox: VoiceVox) {
     this.discordBot = discordBot;
     this.voiceVox = voiceVox;
 
@@ -72,7 +45,7 @@ export default class MessageCreate {
    * messageCreate メイン処理
    */
   public messageCreateEvent(): void {
-    this.discordBot.on('messageCreate', async (message: typeof Message) => {
+    this.discordBot.on('messageCreate', async (message: Message) => {
       reactToUsedMaleEventCommandMessage(message, this.discordBot);
       reactToBirthday235MemberMessage(message, this.discordBot);
       await reactToBirthdayMillionMemberMessage(message, this.discordBot);
@@ -101,830 +74,11 @@ export default class MessageCreate {
       const commandName: string = commandList.shift()!.toLowerCase();
 
       await helpCommand(message, this.discordBot, commandName);
-      await this.birthdayEventCommand(message, commandName, commandList);
-      this.menEventCommand(message, commandName, commandList);
-      this.roomDivisionCommand(this.discordBot, message, commandName);
-      await this.joinVoiceChannelCommand(this.discordBot, message, commandName);
-      await this.disconnectVoiceChannelCommand(message, commandName);
-      this.testCommand(message, commandName, commandList);
+      await birthdayEventCommand(message, this.discordBot, commandName, commandList);
+      await menEventCommand(message, this.discordBot, commandName, commandList);
+      await roomDivisionCommand(this.discordBot, message, commandName);
+      await joinVoiceChannelCommand(message, this.voiceVox, this.discordBot, commandName);
+      await disconnectVoiceChannelCommand(message, this.discordBot, commandName);
     });
-  }
-
-  /**
-   * 235birthdayコマンド 毎月の誕生日祝い企画文章を作成
-   *
-   * @param {Message} message Messageクラス
-   * @param {string} commandName 入力されたコマンド名
-   * @param {string[]} commandList 引数一覧
-   *
-   * @return {void}
-   */
-  private async birthdayEventCommand(
-    message: typeof Message,
-    commandName: string,
-    commandList: string[],
-  ): Promise<void> {
-    if (commandName !== 'birthday' || message.author.id !== this.discordBot.userIdForUtatane)
-      return;
-
-    if (commandList.length < 3 || commandList.length > 4) {
-      message.reply(
-        '235birthdayコマンドを使う場合、birthdayの後にオンライン飲み会を開催したい月、日、時間 （半角数字のみ、曜日は不要） の3つを入力してください。\n任意のテキストを追加したい場合は、3つ入力した後に、追加したいテキストを入力してください。\n※半角スペースで区切るのを忘れずに！！\n\n235birthday 8 15 21',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    let isAllInt: boolean = true;
-
-    for (let i = 0; i < commandList.length; i += 1) {
-      // 追加文章はバリデーションチェックしない
-      if (i === 3) {
-        break;
-      }
-
-      if (!Number.isInteger(Number(commandList[i]))) {
-        isAllInt = false;
-        break;
-      }
-    }
-
-    if (!isAllInt) {
-      message.reply(
-        '半角数字以外が含まれています！\n月、日、時間は全て**半角数字のみ**で入力してください！',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (Number(commandList[0]) < 1 || Number(commandList[0]) > 12) {
-      message.reply('月は1～12の間で入力してください！');
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    const today = new Date();
-    const lastDateTime = new Date(today.getFullYear(), today.getMonth() + 1, 0);
-    // 今月末日を取得
-    const lastDate = lastDateTime.getDate();
-
-    if (Number(commandList[1]) < 1 || Number(commandList[1]) > lastDate) {
-      message.reply(`日は1～${lastDate}の間で入力してください！`);
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (Number(commandList[2]) < 0 || Number(commandList[2]) > 23) {
-      message.reply('時間は0～23の間で入力してください！');
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    const todayYear = today.getFullYear();
-    const eventDateTime = new Date(
-      todayYear,
-      Number(commandList[0]) - 1,
-      Number(commandList[1]),
-      Number(commandList[2]) - 9,
-    );
-    const eventDay = eventDateTime.getDay();
-
-    const week = ['日曜日', '月曜日', '火曜日', '水曜日', '木曜日', '金曜日', '土曜日'];
-
-    const textList = [
-      [
-        `日々のプロデュース業お疲れ様です！！！　${commandList[0]}月に誕生日を迎える方々をご紹介します！！！\n${commandList[0]}月に誕生日を迎えるのは～......\n\n`,
-        `日々のプロデュース業お疲れ様です！${commandList[0]}月にお誕生日を迎える方々のご案内です！\n${commandList[0]}月に誕生日を迎えるのは～…\n\n`,
-        `日々のプロデュース業お疲れ様です！${commandList[0]}月にお誕生日を迎えるメンバーさんの…ご案内です！！\n${commandList[0]}月に誕生日を迎えるのは～…\n\n`,
-        `日々のプロデュース業お疲れ様です！\n${commandList[0]}月期ラウンジオンライン飲み会のご！案！内！です！\n${commandList[0]}月の誕生日は～～～～…\n\n`,
-        `日々のプロデュース業お疲れ様です！${commandList[0]}月に誕生日を迎える方々をご紹介します！\n${commandList[0]}月に誕生日を迎えるのは～…\n\n`,
-        '日々のプロデュース業お疲れ様です！！！今月お誕生日を迎えるのは～…\n\n',
-        `日々のプロデュース業お疲れ様です！${commandList[0]}月が誕生日のメンバーさんをご紹介します！${commandList[0]}月に誕生日を迎えるのは～…\n\n`,
-      ],
-      [
-        `\nです！！！はっぴばーす！と、いうわけで${commandList[0]}月期ラウンジオンライン飲み会のご案内でぇす！！！`,
-        `\nです！はっぴばーす！！！いや～めでたいねぇ（ひなた）\nではでは、${commandList[0]}月期ラウンジオンライン飲み会のご案内です！\n\nQ.ラウンジオンライン飲み会ってなんなん？\nA.ラウンジDiscordに集まってオンラインでやる飲み会だよ！まんまだね！お酒飲めない子はジュースだね！\n　その月の誕生日の人が来たらバースデーを歌ってあげる~~奇習~~お祝いがあるよ！`,
-        `\nです！！！！！おめでとうございますわ～～～～～～～～！！！！！！\nというわけで！${commandList[0]}月期ラウンジオンライン飲み会のご案内です！\n\nQ.ラウンジオンライン飲み会ってなんなん？\nA.ラウンジDiscordに集まってオンラインでやる飲み会だよ！まんまだね！\n　あと、その月の誕生日の人が来たらバースデーを歌ってあげる~~奇習~~お祝いがあるよ！`,
-        '\nです！！！！！！です！おめでとうございます～～～～～～！！！！！！！',
-        '\nです！おめでとうございま～～～す！！！\nというわけで、毎月恒例のオンライン飲み会のご案内です！！！',
-        '\nでぇす！はっぴば～～～す！！！\nということで、月一回、恒例のオンライン飲み会のご案内です！',
-        `\nです！！！おめでとうございま～す！！いやぁめでたいねぇ（ひなた）\nということで${commandList[0]}月期のオンライン飲み会のご案内でーす！`,
-        `\nでーす！はっぴばーす！素敵な一年にしましょうね！\nということで今月もやってきました、${commandList[0]}月期オンライン飲み会のごあんないです！！！`,
-      ],
-      [
-        '遅刻OK早上がりOK、お酒やジュースを飲みながらおしゃべりを楽しむ月一の定例飲み会です！\n皆さんお気軽にご参加お待ちしてま～～～～す(o・∇・o)',
-        '遅れて参加してもOK、眠くなったら先に眠ってもOKの飲み会です！周年イベントが明けても次のイベントはすぐに始まるから（遠い目） お疲れ様会も兼ねて盛り上がってまいりましょう～！多くの皆様方のご参加をお待ちしております！！！！！！！！！お酒お酒お酒お酒！！！！！！！！！',
-        '遅れて参加してもOK!!眠くなったら先に眠ってもOK!!の飲み会です！気持ちアゲていきましょう！！！！ぶいぶい！！！！！！お酒お酒お酒お酒!!!!!!',
-        '遅れて参加してもOK,眠くなったら先に上がってもOKの飲み会です、気ままに楽しみましょう！！！どしどしご参加くださいーーーー！！！！！お酒お酒お酒!!!',
-        '遅れて参加しても良し、眠くなったら先に上がっても寝落ちしてもOKの飲み会です。気軽に和気あいあい楽しみましょう！どしどしご参加くーださい(o・∇・o)',
-        '特に時間などに縛りはございません。好きな時間に来て好きなだけ飲んで話して好きな時間に上がれる飲み会です。まったりのんびり楽しく過ごしましょう～！！！\nお酒お酒お酒お酒お酒!!!!!!!!',
-        '遅刻OK早上がりOK、お酒やジュースを飲みながらおしゃべりを楽しむ月一の定例飲み会です！皆さんお気軽にご参加お待ちしてま~~~~す',
-      ],
-    ];
-
-    let eventText: string = `@everyone\n${textList[0][Math.floor(Math.random() * textList[0].length)]}`;
-
-    BirthdayFor235MemberRepository.getThisMonthBirthdayMember(commandList[0]).then(
-      (
-        birthdayMemberList: {
-          name: string;
-          user_id: string;
-          month: number;
-          date: number;
-          speaker_id: number;
-        }[],
-      ) => {
-        birthdayMemberList.forEach((birthdayMember) => {
-          eventText += `**${birthdayMember.date}日...${birthdayMember.name}さん**\n`;
-        });
-
-        eventText += textList[1][Math.floor(Math.random() * textList[1].length)];
-
-        eventText += `\n\n**開催日：${commandList[0]}月${commandList[1]}日 （${week[eventDay]}）**\n**時間：${commandList[2]}時ごろ～眠くなるまで**\n**場所：ラウンジDiscord雑談通話**\n**持参品：**:shaved_ice::icecream::ice_cream::cup_with_straw::champagne_glass::pizza::cookie:\n\n`;
-
-        eventText += textList[2][Math.floor(Math.random() * textList[2].length)];
-
-        if (commandList.length === 4) {
-          eventText += `\n${commandList[3]}`;
-        }
-
-        message.channel.send(eventText);
-      },
-    );
-
-    const description = commandList[3] ?? 'ラウンジDiscordに集まってオンラインでやる飲み会だよ！';
-
-    const eventId = await this.createScheduleEvent(
-      `${commandList[0]}月期ラウンジオンライン飲み会`,
-      description,
-      eventDateTime,
-      this.discordBot.voiceChannelIdFor235ChatPlace,
-    );
-
-    setTimeout(
-      () =>
-        message.reply(
-          `うたたねさん、今回もお疲れ様です！\n\n【${commandList[0]}月期ラウンジオンライン飲み会】のイベントを作成しました！\nhttps://discord.com/events/${this.discordBot.serverIdFor235}/${eventId}\n\nいつもありがとうございます♪`,
-        ),
-      6_000,
-    );
-
-    setTimeout(() => {
-      message
-        .delete()
-        .then(() => console.log('message deleting.'))
-        .catch(() => console.log('message is deleted.'));
-    }, this.setTimeoutSec);
-  }
-
-  /**
-   * 235menコマンド 男子会の日程を決める文章を作成
-   *
-   * @param {Message} message Messageクラス
-   * @param {string} commandName 入力されたコマンド名
-   * @param {string[]} commandList 引数一覧
-   *
-   * @return {void}
-   */
-  private menEventCommand(
-    message: typeof Message,
-    commandName: string,
-    commandList: string[],
-  ): void {
-    if (commandName !== 'men' || message.author.id !== this.discordBot.userIdForUtatane) return;
-
-    if (commandList.length < 1 || commandList.length > 10) {
-      message.reply(
-        '235menコマンドは、235士官学校の日程を決めるために使用するコマンドです。\n開校したい日程を**半角スペースで区切って**入力してください。（半角数字のみ、月、曜日などは不要）\n入力できる日程の数は**2～10個まで**です！\n\n235men 8 12 15 21',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    const isAllInt: boolean = commandList.every((date) => Number.isInteger(Number(date)));
-
-    if (!isAllInt) {
-      message.reply('半角数字以外が含まれています！\n日程は**半角数字のみ**で入力してください！');
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (MessageCreate.isExistsSameValue(commandList)) {
-      message.reply('同じ日程が入力されています！');
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    const today = new Date();
-    const lastDateTime = new Date(today.getFullYear(), today.getMonth() + 1, 0);
-    // 今月末日を取得
-    const lastDate = lastDateTime.getDate();
-
-    const isValidDate: boolean = commandList.every(
-      (date) => Number(date) >= 1 && Number(date) <= lastDate,
-    );
-
-    if (!isValidDate) {
-      message.reply(`日は1～${lastDate}の間で入力してください！`);
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    commandList.sort((a, b) => Number(a) - Number(b));
-
-    const todayYear = today.getFullYear();
-    const todayMonth = today.getMonth() + 1;
-    const eventDayList = commandList.map((date) =>
-      new Date(todayYear, todayMonth - 1, Number(date)).getDay(),
-    );
-
-    const week = ['日曜日', '月曜日', '火曜日', '水曜日', '木曜日', '金曜日', '土曜日'];
-
-    const textList = [
-      [
-        `ふみこ男子の皆様方～～～～～～～～～～～！${todayMonth}月期の235士官学校開校日を決めたいと思いますわ～～～～～！！！日程なんですけど、\n\n`,
-        `ふみこ男子の皆様方～～～～～～～～～！${todayMonth}月期の235士官学校開校日を決めたいと思います！その日程なんですけど、\n\n`,
-      ],
-      [
-        '\n誠に勝手ながらこのいずれかの日程でやろうと思いますので、スタンプで反応を頂けると嬉しいです～～～～ふみこ男子の皆様方！よろしくおねがいしますわね！！！！！！！！！ﾍｹｯ!!!!!!!!',
-        '\n真に勝手ながらこのいずれかにしようと思いますので、2~3日中にスタンプで反応を頂けると幸いです！よろしくお願いしま～～～～～～～す🙏',
-      ],
-    ];
-
-    let eventText: string = `@everyone\n${textList[0][Math.floor(Math.random() * textList[0].length)]}`;
-
-    for (let i = 0; i < commandList.length; i += 1) {
-      eventText += `**${todayMonth}月${commandList[i]}日 （${week[eventDayList[i]]}）…　${this.maleEventEmojiList[i]}**\n`;
-    }
-
-    eventText += textList[1][Math.floor(Math.random() * textList[1].length)];
-
-    message.channel.send(eventText);
-
-    this.discordBot.usedMaleEventCommandReactionCount = commandList.length;
-
-    setTimeout(
-      () => message.reply('うたたねさん、今回もお疲れ様です！\nいつもありがとうございます♪'),
-      6_000,
-    );
-
-    setTimeout(() => {
-      message
-        .delete()
-        .then(() => console.log('message deleting.'))
-        .catch(() => console.log('message is deleted.'));
-    }, this.setTimeoutSec);
-  }
-
-  /**
-   * 235roomdivisionコマンド ボイスチャンネルに参加しているメンバーを分ける
-   *
-   * @param {Client} client Clientクラス
-   * @param {Message} message Messageクラス
-   * @param {string} commandName 入力されたコマンド名
-   *
-   * @return {void}
-   */
-  private roomDivisionCommand(
-    client: typeof Client,
-    message: typeof Message,
-    commandName: string,
-  ): void {
-    if (commandName !== 'roomdivision') return;
-
-    let participatingVoiceChannelMemberList: {
-      userName: string;
-      userId: string;
-    }[] = client.voice.client.channels.cache
-      .get(this.discordBot.voiceChannelIdFor235ChatPlace)
-      .members.filter((member: typeof GuildMember) => member.user.bot === false)
-      .map((member: typeof GuildMember) => {
-        return {
-          userName: member.user.globalName,
-          userId: member.user.id,
-        };
-      });
-
-    const isParticipateVoiceChannelUsedCommandMember: boolean =
-      participatingVoiceChannelMemberList.some(
-        (participatingVoiceChannelMember: { userName: string; userId: string }) =>
-          participatingVoiceChannelMember.userId === message.author.id,
-      );
-
-    if (!isParticipateVoiceChannelUsedCommandMember) {
-      message.reply(
-        '235roomdivision コマンドは、【雑談１】ボイスチャンネルに参加しているメンバーが使用できるコマンドです。',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (participatingVoiceChannelMemberList.length < 10) {
-      message.reply(
-        '【雑談１】ボイスチャンネルに参加しているメンバーの人数が10人未満のため、分けることが出来ません！',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    message.channel.sendTyping();
-
-    let halfIndex: number = 0;
-    let halfIndex2: number = 0;
-
-    // 配列を分ける
-    if (participatingVoiceChannelMemberList.length % 2 === 0) {
-      halfIndex = Math.floor(participatingVoiceChannelMemberList.length / 2) - 1;
-      halfIndex2 = participatingVoiceChannelMemberList.length - halfIndex - 1;
-    } else {
-      halfIndex = Math.floor(participatingVoiceChannelMemberList.length / 2);
-      halfIndex2 = participatingVoiceChannelMemberList.length - halfIndex;
-    }
-
-    let duplicationCount: number = 100;
-    let halfMemberList: { userName: string; userId: string }[] = [];
-    let halfMemberList2: { userName: string; userId: string }[] = [];
-
-    while (duplicationCount >= 3) {
-      // 初期化
-      participatingVoiceChannelMemberList = MessageCreate.shuffle(
-        participatingVoiceChannelMemberList,
-      );
-
-      halfMemberList = [];
-      halfMemberList2 = [];
-
-      for (let i = 0; i <= halfIndex; i += 1) {
-        halfMemberList.push(participatingVoiceChannelMemberList[i]);
-      }
-
-      for (let i = halfIndex2; i < participatingVoiceChannelMemberList.length; i += 1) {
-        halfMemberList2.push(participatingVoiceChannelMemberList[i]);
-      }
-
-      // 3人以上被ってないかチェック
-      duplicationCount = halfMemberList2.filter(
-        (participatingVoiceChannelMember: { userName: string; userId: string }) => {
-          return (
-            this.discordBot.dividedUserIdList.indexOf(participatingVoiceChannelMember.userId) !== -1
-          );
-        },
-      ).length;
-    }
-
-    // 2個目の配列のメンバーを雑談２ボイスチャンネルに移動
-    const dividedUserNameList = halfMemberList.map(
-      (dividedMember: { userName: string; userId: string }) => dividedMember.userName,
-    );
-
-    const dividedUserNameList2 = halfMemberList2.map(
-      (dividedMember: { userName: string; userId: string }) => dividedMember.userName,
-    );
-
-    const dividedUserIdList2 = halfMemberList2.map(
-      (dividedMember: { userName: string; userId: string }) => dividedMember.userId,
-    );
-
-    this.discordBot.dividedUserIdList = dividedUserIdList2;
-
-    setTimeout(
-      () =>
-        message.reply(
-          `このような結果になりました！\n\n**【雑談１】**\n------------------------------------------------------------\n${dividedUserNameList.join('\n')}\n------------------------------------------------------------\n\n**【雑談２】**\n------------------------------------------------------------\n${dividedUserNameList2.join('\n')}\n------------------------------------------------------------\n\n自動で分けられますのでしばらくお待ちください。`,
-        ),
-      2_000,
-    );
-
-    setTimeout(() => {
-      let divisionCount: number = 0;
-
-      const roomDivideTimer = setInterval(() => {
-        switch (divisionCount) {
-          case halfMemberList2.length:
-            message
-              .delete()
-              .then(() => console.log('message deleting.'))
-              .catch(() => console.log('message is deleted.'));
-            clearInterval(roomDivideTimer);
-            break;
-
-          default:
-            client.guilds.cache
-              .get(this.discordBot.serverIdFor235)
-              .members.fetch(dividedUserIdList2[divisionCount])
-              .then((member: typeof GuildMember) =>
-                member.voice.setChannel(this.discordBot.voiceChannelIdFor235ChatPlace2),
-              );
-
-            divisionCount += 1;
-            break;
-        }
-      }, 1_000);
-    }, 9_000);
-  }
-
-  /**
-   * 235joinコマンド コマンドを入力したメンバーが入っているボイスチャンネルに参加
-   *
-   * @param {Client} client Clientクラス
-   * @param {Message} message Messageクラス
-   * @param {string} commandName 入力されたコマンド名
-   *
-   * @return {void}
-   */
-  private async joinVoiceChannelCommand(
-    client: typeof Client,
-    message: typeof Message,
-    commandName: string,
-  ) {
-    if (commandName !== 'join') return;
-
-    const usedCommandMember = await message.guild.members.fetch(message.author.id);
-    const memberJoinVoiceChannel = usedCommandMember.voice.channel;
-
-    if (
-      client.connection !== undefined &&
-      client.connection.joinConfig.channelId === memberJoinVoiceChannel.id
-    ) {
-      const embed = new EmbedBuilder()
-        .setTitle('既に接続されています！')
-        .setFields({ name: 'ボイスチャンネル名', value: memberJoinVoiceChannel.name })
-        .setColor('#FF0000')
-        .setTimestamp();
-
-      message.reply({ embeds: [embed] });
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (memberJoinVoiceChannel === null) {
-      message.reply(
-        '235joinコマンドを使用することで、使用したメンバーが参加しているボイスチャンネルに235botが参加して、そのボイスチャンネルの聞き専チャンネルに投稿されたテキストを読み上げます！\nボイスチャンネルに参加してから再度このスラッシュコマンドを使用していただくか、もしくはテキストで「235join」と入力していただければボイスチャンネルに参加します！',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (!memberJoinVoiceChannel.joinable || !memberJoinVoiceChannel.speakable) {
-      message.reply(
-        '参加先のボイスチャンネルに接続できなかったか、もしくは参加先のボイスチャンネルで音声を再生する権限がありませんでした；；',
-      );
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    this.discordBot.connection = joinVoiceChannel({
-      channelId: memberJoinVoiceChannel.id,
-      guildId: message.guild.id,
-      adapterCreator: message.guild.voiceAdapterCreator,
-      selfMute: false,
-      selfDeaf: true,
-    });
-
-    this.discordBot.connection.subscribe(this.discordBot.audioPlayer);
-
-    const connectVoice =
-      client.connectVoiceList[Math.floor(Math.random() * client.connectVoiceList.length)];
-
-    const filePath = './data/voice';
-    const wavFile = `${filePath}/${usedCommandMember.user.id}.wav`;
-
-    if (!fs.existsSync(filePath)) fs.mkdirSync(filePath, { recursive: true });
-
-    await VoiceVox.generateAudioFile(connectVoice, wavFile, client.speakerId);
-
-    this.voiceVox.addWavFileToQueue(wavFile);
-
-    const embed = new EmbedBuilder()
-      .setTitle('接続されました！')
-      .setFields({ name: 'ボイスチャンネル名', value: memberJoinVoiceChannel.name })
-      .setColor('#00FF99')
-      .setTimestamp();
-
-    message.reply({ embeds: [embed] });
-
-    setTimeout(() => {
-      message
-        .delete()
-        .then(() => console.log('message deleting.'))
-        .catch(() => console.log('message is deleted.'));
-    }, this.setTimeoutSec);
-  }
-
-  /**
-   * 235disconnectコマンド 235botをボイスチャンネルから切断
-   *
-   * @param {Message} message Messageクラス
-   * @param {string} commandName 入力されたコマンド名
-   *
-   * @return {void}
-   */
-  private async disconnectVoiceChannelCommand(message: typeof Message, commandName: string) {
-    if (commandName !== 'disconnect') return;
-
-    const usedCommandMember = await message.guild.members.fetch(message.author.id);
-    const memberJoinVoiceChannel = usedCommandMember.voice.channel;
-
-    if (this.discordBot.connection === undefined) {
-      const embed = new EmbedBuilder()
-        .setTitle('まだボイスチャンネルに接続されていません！')
-        .setColor('#FF0000')
-        .setTimestamp();
-
-      message.reply({ embeds: [embed] });
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    if (
-      memberJoinVoiceChannel === null ||
-      this.discordBot.connection.joinConfig.channelId !== memberJoinVoiceChannel.id
-    ) {
-      const embed = new EmbedBuilder()
-        .setTitle(
-          '切断できるのは235botが入っているボイスチャンネルに参加しているメンバーだけです！',
-        )
-        .setColor('#FFCC00')
-        .setTimestamp();
-
-      message.reply({ embeds: [embed] });
-
-      setTimeout(() => {
-        message
-          .delete()
-          .then(() => console.log('message deleting.'))
-          .catch(() => console.log('message is deleted.'));
-      }, this.setTimeoutSec);
-
-      return;
-    }
-
-    this.discordBot.connection.destroy();
-    this.discordBot.connection = undefined;
-
-    const embed = new EmbedBuilder()
-      .setTitle('切断されました！')
-      .setFields({ name: 'ボイスチャンネル名', value: memberJoinVoiceChannel.name })
-      .setColor('#00FF99')
-      .setTimestamp();
-
-    message.reply({ embeds: [embed] });
-
-    setTimeout(() => {
-      message
-        .delete()
-        .then(() => console.log('message deleting.'))
-        .catch(() => console.log('message is deleted.'));
-    }, this.setTimeoutSec);
-  }
-
-  /**
-   * 235testコマンド 新しい機能を追加する時に実験とかする用
-   *
-   * @param {Message} message Messageクラス
-   * @param {string} commandName 入力されたコマンド名
-   * @param {string[]} commandList 引数一覧
-   *
-   * @return {void}
-   */
-  private testCommand(message: typeof Message, commandName: string, commandList: string[]): void {
-    if (
-      commandName !== 'test' ||
-      message.author.id !== this.discordBot.userIdForMaki ||
-      commandList.length === 0
-    )
-      return;
-
-    switch (commandList[0]) {
-      case 'birthday_for_235_members':
-        BirthdayFor235Member.findAll({ raw: true }).then(
-          (
-            member: {
-              name: string;
-              user_id: string;
-              month: number;
-              date: number;
-            }[],
-          ) => console.log(member),
-        );
-
-        break;
-      case 'birthday_for_million_members':
-        BirthdayForMillionMember.findAll({ raw: true }).then(
-          (
-            member: {
-              name: string;
-              month: number;
-              date: number;
-              img: string;
-            }[],
-          ) => console.log(member),
-        );
-
-        break;
-      case 'delete_messages':
-        DeleteMessage.findAll({ raw: true }).then(
-          (
-            data: {
-              message_id: string;
-              date: number;
-            }[],
-          ) => console.log(data),
-        );
-
-        break;
-      case 'dict_words':
-        DictWord.findAll({ raw: true }).then(
-          (
-            data: {
-              word: string;
-              how_to_read: string;
-            }[],
-          ) => console.log(data),
-        );
-
-        break;
-    }
-
-    setTimeout(() => {
-      message
-        .delete()
-        .then(() => console.log('message deleting.'))
-        .catch(() => console.log('message is deleted.'));
-    }, this.setTimeoutSec);
-  }
-
-  /**
-   * オンライン飲み会、男子会のイベントを作成
-   *
-   * @param {string} eventTitle 作成するイベント名
-   * @param {string} description 作成するイベントの説明
-   * @param {Date} scheduledStartTime イベントの開始日
-   * @param {string} channelId 開始場所（ボイスチャンネルのID）
-   * @param {string | null} image アイコン画像
-   *
-   * @return {string}
-   */
-  private async createScheduleEvent(
-    eventTitle: string,
-    description: string,
-    scheduledStartTime: Date,
-    channelId: string,
-    image: string | null = null,
-  ): Promise<string> {
-    const guild = this.discordBot.guilds.cache.get(this.discordBot.serverIdFor235);
-    const eventManager = new GuildScheduledEventManager(guild);
-
-    const options = {
-      name: eventTitle,
-      scheduledStartTime,
-      privacyLevel: GuildScheduledEventPrivacyLevel.GuildOnly,
-      entityType: GuildScheduledEventEntityType.Voice,
-      description,
-      channel: channelId,
-      image,
-    };
-
-    const createdEvent = await eventManager.create(options);
-
-    return createdEvent.id;
-  }
-
-  /**
-   * 指定された配列の中に重複している要素があるかチェック
-   *
-   * @param {string[]} targetList 配列
-   *
-   * @return {boolean}
-   */
-  private static isExistsSameValue(targetList: string[]): boolean {
-    const set = new Set(targetList);
-
-    return set.size !== targetList.length;
-  }
-
-  /**
-   * 配列の要素の中身をシャッフル
-   *
-   * @param {T[]} memberList シャッフル対象のメンバーリスト
-   *
-   * @return {T[]}
-   */
-  private static shuffle<T>(memberList: T[]): T[] {
-    const out = Array.from(memberList);
-
-    for (let i = out.length - 1; i > 0; i -= 1) {
-      const r = Math.floor(Math.random() * (i + 1));
-      const tmp = out[i];
-      out[i] = out[r];
-      out[r] = tmp;
-    }
-
-    return out;
   }
 }
